demoProject/requirements.txt
asgiref==3.7.2
Django==4.2 # 4.2 5.0.2
django-filter==24.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Markdown==3.5.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
